Move TokenAuthority deployment to TokenClient
